move timeseries methods to series
